Add files via upload
Story:
You and your crew are stuck in a room. You just have one attempt to unlock and escape. If you fail you are locked forever. The key to the door will unlock only when the CLOCK shows 12:00:00.
Unfortunately, there is no clock in the room which can show you the time. However, you have a computer!
Create a program that draws the clock for you and shows the time to find your way out.